// isem-finanzas-web/src/app/@finanzas/@suficiencias/pages/models/observacion.model.ts
import { NgbDate } from '@ng-bootstrap/ng-bootstrap';

export class ObservacionModel {
  id_observacion?: number;
  id_auditoria?: number;
  id_requerimiento?: number;
  id_suficiencia?: number;
  observacion?: string;
  fecha_creacion?: string;
}
export class Observacio {


    id_observacion?: number;
    id_auditoria?: number;
    id_suficiencia?: number;
    observacion?: string;

}